requirements.txt
pandas
numpy
requests
pymongo
dnspython
python-dotenv
scikit-learn
keras
tensorflow
matplotlib
seaborn
zenml
streamlit
mlflow
zenml[server]
pymongo[srv]
# -e .